Refactor component structure by moving CircularProgress and TaskSection to sub-components
- Deleted the CircularProgress and TasksTable components to streamline the codebase.
- Updated import paths in FocusMetrics and MyTasks components to reflect the new sub-component structure.
- Enhanced organization of components for better maintainability and clarity.
