+++ src/components/my-tasks.tsx
@@ -10,7 +10,7 @@ import {
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { tasks } from "@/data/dummydata"
-import { TaskSection } from "./my-tasks-sub-components/task-section"
+import { TaskSection } from "./sub-components/task-section"
